requirements.txt
streamlit==1.25.0
streamlit-webrtc==1.0.1
Pillow==9.5.0
numpy==1.26.4
